+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 * Information dialog on device's services
 * Compose vNotes
+* --delete option for blueman-sendto
 
 ### Changes